Matching Game
When the game starts, three faces are shown on the left and two are
shown on the right. The left and right sides are identical, except for
one thing: the left side has one extra face. The user needs to click on
that extra face. If anything except the correct face is clicked, a
message is displayed saying that the game is over. If the correct face
is clicked, all the currently displayed faces are deleted and a new set
of faces is shown at random positions. Each time a new set of faces is
shown there will be 3 more faces than before, on both the left and the
right sides. There will always be one extra face shown on the left